backend/requirements.txt
flask #lightweight Python web framework
psycopg2-binary #PostgreSQL connector for Python
flask-cors
Flask-JWT-Extended
Flask-Bcrypt